+++ elm/src/ZkInterface.elm
@@ -22,7 +22,6 @@ type SendMsg
     | SaveImportZkNotes (List Data.ImportZkNote)
     | PowerDelete S.TagSearch
     | SetHomeNote Int
-    | Yeet Data.Yeet
 
 
 type ServerResponse
@@ -185,12 +184,6 @@ encodeSendMsg sm =
             JE.object
                 [ ( "what", JE.string "sethomenote" )
                 , ( "data", JE.int id )
-                ]
-
-        Yeet yt ->
-            JE.object
-                [ ( "what", JE.string "yeet" )
-                , ( "data", Data.encodeYeet yt )
                 ]
