+++ server/Cargo.toml
@@ -37,5 +37,4 @@ clap = "2.33.2"
 timer = "0.2.0"
 chrono = "0.4.15"
 either = "1.6.1"
-glob = "0.3.0"
 sha256 = "1.1.1"
