+++ zkprotocol/src/content.rs
@@ -170,9 +170,3 @@ pub struct ZkNoteEditWhat {
   pub what: String,
   pub zne: ZkNoteEdit,
 }
-
-#[derive(Deserialize, Debug)]
-pub struct Yeet {
-  pub url: String,
-  pub audio: bool,
-}
